Add feature gate for WinPTY
