+++ alacritty/Cargo.toml
@@ -11,7 +11,7 @@ edition = "2018"
 [dependencies]
 alacritty_terminal = { path = "../alacritty_terminal", default-features = false }
 clap = "2"
-log = "0.4"
+log = { version = "0.4", features = ["std"] }
 time = "0.1.40"
 fnv = "1"
 serde = { version = "1", features = ["derive"] }
@@ -47,9 +47,10 @@ winapi = { version = "0.3.7", features = ["impl-default", "wincon"]}
 embed-resource = "1.3"
 
 [features]
-default = ["wayland", "x11"]
+default = ["wayland", "x11", "winpty"]
 x11 = ["alacritty_terminal/x11"]
 wayland = ["alacritty_terminal/wayland"]
+winpty = ["alacritty_terminal/winpty"]
 # Enabling this feature makes shaders automatically reload when changed
 live-shader-reload = []
 nightly = []
